data/scrap/requirements.txt
requests
pandas
beautifulsoup4
psycopg2-binary==2.9.10
selenium
webdriver-manager